message_broker/pyproject.toml
[project]
name = "message_broker"
version = "0.1.0"
description = "Modular, pluggable async message broker with unified interfaces for Redis, Kafka, RabbitMQ, and in-memory transports"
readme = "Readme.md"
requires-python = ">=3.10"
authors = [ { name = "LYIK" } ]

dependencies = [
    "redis[asyncio]>=5.0.0",
    "aiokafka>=0.10.0",
    "pydantic>=2.0.0",
    "structlog>=24.1.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
# Install with: pip install message_broker[otel]
otel = [
    "opentelemetry-api>=1.20.0"
]

[project.scripts]
mb-messager = "message_broker.cli_messager:run"
mb-receiver = "message_broker.cli_receiver:run"

[project.entry-points."message_broker.adapters"]
redis = "message_broker.adapters.redis:redis_factory"
kafka = "message_broker.adapters.kafka:kafka_factory"

[tool.setuptools]
packages = [
    "message_broker", 
    "message_broker.src", 
    "message_broker.src.core", 
    "message_broker.src.adapters"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"